+++ scif/client/test.py
@@ -1,8 +1,8 @@
 '''
 
-Copyright (C) 2017-2018 The Board of Trustees of the Leland Stanford Junior
+Copyright (C) 2017 The Board of Trustees of the Leland Stanford Junior
 University.
-Copyright (C) 2017-2018 Vanessa Sochat.
+Copyright (C) 2016-2017 Vanessa Sochat.
 
 This program is free software: you can redistribute it and/or modify it
 under the terms of the GNU Affero General Public License as published by
@@ -21,31 +21,25 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 from scif.logger import bot
 import sys
+import pwd
 import os
-
-    
 
 def main(args,parser,subparser):
 
     from scif.main import ScifRecipe
-    client = ScifRecipe(writable=False, quiet=True)
-    longlist = args.longlist
-    
-    result = []
-    for app in client.apps():
-        config = client.get_appenv(app)
+    cmd = args.cmd
+    client = ScifRecipe(quiet=True, writable=args.writable)
 
-        # Long listing includes a number with path to app
-        if longlist is True:
-            result.append([app.rjust(10), config['SCIF_APPROOT']]) 
-        else:
-            result.append(app.rjust(10))
+    if len(cmd) == 0:
+        bot.warning('You must supply an appname to test.')
+        bot.custom(prefix="Example: ", message="scif test <app>")
+        sys.exit(1)
 
-    if len(result) > 0:
+    app = cmd.pop(0)
 
-        if longlist is True:
-            header = "[app]              [root]"
-            bot.custom(prefix="SCIF", message=header, color="CYAN")
-            bot.table(result)
-        else:
-            print('\n'.join(result))
+    # Remaining arguments indicate options/args to pass on
+    if len(cmd) == 0:
+        cmd = None
+
+    client = ScifRecipe(quiet=True, writable=args.writable)
+    client.test(app, args=cmd)