+++ scif/main/parser.py
@@ -74,7 +74,7 @@ def load_recipe(path):
         # Read in spec, skip empty lines, don't strip remaining
         spec = [x.strip('\n') 
                for x in read_file(path) 
-               if not re.match('^\s+$', x.strip('\n'))]
+               if not re.match(r'^\s+$', x.strip('\n'))]
 
         spec = [x for x in spec if x not in ['', None]]
         config = OrderedDict()
@@ -101,7 +101,7 @@ def load_recipe(path):
                 parts = line.split(' ')
                 if len(parts) > 1:
                     name = ' '.join(parts[1:])          
-                section = re.sub('[%]|(\s+)','',parts[0]).lower()
+                section = re.sub(r'[%]|(\s+)','',parts[0]).lower()
                 config = add_section(config=config,
                                      section=section,
                                      name=name)
@@ -152,7 +152,11 @@ def finish_recipe(config, global_section='apps'):
         if "appenv" in config[global_section][app]:
             appenv = config[global_section][app]['appenv']
 
-            # If runscript is defined, add source to first line
+            # If runscript or test is defined, add source to first line      
+            if "apptest" in config[global_section][app]:
+                apptest = config[global_section][app]['apptest']
+                config[global_section][app]['apptest'] =  appenv + apptest
+
             if "apprun" in config[global_section][app]:
                 apprun = config[global_section][app]['apprun']
                 config[global_section][app]['apprun'] =  appenv + apprun