+++ scif/tests/test_run.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+CONTAINER_NAME=${1:-}
+
+echo "Testing run of each app:"
+docker run -it "${CONTAINER_NAME}" run hello-world-echo
+docker run -it "${CONTAINER_NAME}" run hello-world-script
+docker run -it "${CONTAINER_NAME}" run hello-world-env