+++ CHANGELOG.md
@@ -14,7 +14,8 @@ represented by the pull requests that fixed them.
 
 Versions in parentheses coincide with what is available on [pypi](https://pypi.org/project/scif/).
 
-## [xxx](https://github.com/vsoch/scif/tree/master) (development)
+## [xxx](https://github.com/vsoch/scif/tree/master) (master)
+ - adding test to client, along with tests for functions (and on CircleCI) (0.0.74)
  - ensuring that other application environment variables available at install (0.0.73)
  - adding custom shell command to client, and ensuring paths get appended (0.0.72)
  - allowing support for periods in app names (0.0.71)